Add path_prefix option
Godot will generate paths relative to res:// however if the project is
not in the root if the git repository the path in the licenses.md file
won't be correct.

The path_prefix option allows the user to specify the equivalent path to
res:// in the git repo.

Defaults to "/".